Add test cases for FrozenTrial.validate
