+++ export_test.go
@@ -1,0 +1,5 @@
+package goptuna
+
+var (
+	ExportFrozenTrialValidate = (*FrozenTrial).validate
+)
